@@ -5,23 +5,23 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@types/history": "4.7.2",
-    "@types/node": "^12.7.2",
-    "@types/react": "^16.9.2",
-    "@types/react-dom": "^16.8.5",
+    "@types/history": "4.7.3",
+    "@types/node": "^12.7.9",
+    "@types/react": "^16.9.4",
+    "@types/react-dom": "^16.9.1",
     "@types/react-hot-loader": "4.1.0",
-    "@types/react-redux": "^7.1.2",
-    "@types/react-router": "^5.0.3",
-    "@types/react-router-dom": "^4.3.4",
+    "@types/react-redux": "^7.1.4",
+    "@types/react-router": "^5.1.1",
+    "@types/react-router-dom": "^5.1.0",
     "@types/react-router-redux": "^5.0.18",
-    "@types/webpack": "^4.32.1",
+    "@types/webpack": "^4.39.2",
     "@types/webpack-env": "^1.14.0",
     "aspnet-webpack": "3.0.0",
     "aspnet-webpack-react": "4.0.0",
     "awesome-typescript-loader": "5.2.1",
     "css-hot-loader": "^1.4.4",
     "css-loader": "^3.2.0",
-    "event-source-polyfill": "^1.0.8",
+    "event-source-polyfill": "^1.0.9",
     "file-loader": "4.2.0",
     "isomorphic-style-loader": "^5.1.0",
     "mini-css-extract-plugin": "^0.8.0",
@@ -31,45 +31,45 @@
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
     "postcss-safe-parser": "^4.0.1",
-    "react-dev-utils": "^9.0.3",
-    "react-hot-loader": "^4.12.11",
-    "sass-loader": "7.2.0",
+    "react-dev-utils": "^9.1.0",
+    "react-hot-loader": "^4.12.14",
+    "sass-loader": "8.0.0",
     "sass-resources-loader": "^2.0.1",
     "style-loader": "1.0.0",
-    "terser-webpack-plugin": "^1.4.1",
+    "terser-webpack-plugin": "^2.1.2",
     "tslib": "1.10.0",
-    "tslint": "^5.18.0",
+    "tslint": "^5.20.0",
     "tslint-loader": "3.6.0",
-    "typescript": "^3.5.3",
+    "typescript": "^3.6.3",
     "url-loader": "2.1.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.6",
-    "webpack-dev-middleware": "^3.7.0",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-merge": "^4.2.1"
+    "webpack-merge": "^4.2.2"
   },
   "dependencies": {
     "@aspnet/signalr": "^1.1.4",
-    "@fortawesome/fontawesome-svg-core": "^1.2.21",
-    "@fortawesome/free-solid-svg-icons": "^5.10.1",
-    "@fortawesome/react-fontawesome": "^0.1.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@fortawesome/react-fontawesome": "^0.1.5",
     "aspnet-prerendering": "^3.0.1",
     "autobind-decorator": "^2.4.0",
     "axios": "^0.19.0",
     "bulma": "^0.7.5",
     "connected-react-router": "^6.5.2",
     "domain-task": "^3.0.3",
-    "history": "^4.9.0",
+    "history": "^4.10.1",
     "node-noop": "^1.0.0",
     "promise-polyfill": "^8.1.3",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-redux": "^7.1.0",
-    "react-router-dom": "^5.0.1",
+    "react": "^16.10.1",
+    "react-dom": "^16.10.1",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.1.2",
     "react-router-redux": "^4.0.8",
-    "react-toastify": "^5.3.2",
+    "react-toastify": "^5.4.0",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
-    "styled-components": "^4.3.2"
+    "styled-components": "^4.4.0"
   }
 }
