@@ -9,23 +9,30 @@ type CheckboxProps = {
   onCheck: (checked: boolean) => void;
 };
 
-const Checkbox: React.FC<CheckboxProps> = props => {
+const Checkbox: React.FC<CheckboxProps> = ({
+  onCheck,
+  checked,
+  disabled,
+  wrapperClass,
+  controlClass,
+  trailingLabel,
+}) => {
   const handleOnCheck: React.MouseEventHandler<HTMLInputElement> = e => {
-    props.onCheck(!!e.currentTarget.checked);
+    onCheck(!!e.currentTarget.checked);
   };
 
   return (
-    <div className={`control ${props.wrapperClass || ''}`}>
-      <p className={`checkbox-control ${props.controlClass || ''}`}>
+    <div className={(`control ${wrapperClass || ''}`).trim()}>
+      <p className={(`checkbox-control ${controlClass || ''}`).trim()}>
         <label>
           <input
             type='checkbox'
-            disabled={props.disabled}
-            defaultChecked={!!props.checked}
+            disabled={disabled}
+            defaultChecked={!!checked}
             onClick={handleOnCheck}
           />
           <i className='helper' />
-          {props.trailingLabel && <span>{props.trailingLabel}</span>}
+          {trailingLabel && <span>{trailingLabel}</span>}
         </label>
       </p>
     </div>
@@ -33,35 +40,3 @@ const Checkbox: React.FC<CheckboxProps> = props => {
 };
 
 export default Checkbox;
-
-/**
- * ORIGINAL CLASS IMPLEMENTATION
- */
-
-//export default class Checkbox extends React.PureComponent<CheckboxProps> {
-//    static defaultProps = {
-//        wrapperClass: '',
-//        controlClass: ''
-//    };
-
-//    public render(): React.ReactNode {
-//        return (
-//            <div className={`control ${this.props.wrapperClass}`}>
-//                <p className={`checkbox-control ${this.props.controlClass}`}>
-//                    <label>
-//                        <input type='checkbox'
-//                               disabled={this.props.disabled}
-//                               defaultChecked={!!this.props.checked}
-//                               onClick={this.handleCheckboxClick} />
-//                        <i className='helper'></i>
-//                        { this.props.trailingLabel && <span>{this.props.trailingLabel}</span> }
-//                    </label>
-//                </p>
-//            </div>
-//        );
-//    }
-
-//    private handleCheckboxClick: React.MouseEventHandler<HTMLInputElement> = (e) => {
-//        this.props.dispatchHandler(!!e.currentTarget.checked);
-//    }
-//}
