@@ -7,24 +7,30 @@ import { DROPDOWN_TEST_DATA } from '../config/constants';
 import { Checkbox, Dropdown } from '../components/controls';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-type FormProps = ReturnType<typeof reducer> &
-  typeof actionCreators &
-  RouteComponentProps<{}>;
+type FormProps = ReturnType<typeof reducer> & typeof actionCreators & RouteComponentProps<{}>;
 
-const Form: React.FC<FormProps> = props => {
+const Form: React.FC<FormProps> = ({
+  count,
+  doCheck,
+  increment,
+  decrement,
+  selectOption,
+  checkboxValue,
+  selectedDropdownOption,
+}) => {
   const renderCounterGroup: React.ReactNode = (
     <div className='column'>
       <h3 className='title is-4'>Counter</h3>
       <h5 className='subtitle is-5'>Simple example of a React component</h5>
       <p className='subtitle is-5'>
-        Current count: <strong>{props.count}</strong>
+        Current count: <strong>{count}</strong>
       </p>
       <p className='buttons incrementer-buttons'>
-        <button className='button is-danger' onClick={() => props.decrement()}>
+        <button className='button is-danger' onClick={() => decrement()}>
           <FontAwesomeIcon icon='minus' />
           Decrement
         </button>
-        <button className='button is-success' onClick={() => props.increment()}>
+        <button className='button is-success' onClick={() => increment()}>
           <FontAwesomeIcon icon='plus' />
           Increment
         </button>
@@ -37,15 +43,15 @@ const Form: React.FC<FormProps> = props => {
       <h3 className='title is-4'>Dropdown</h3>
       <h5 className='subtitle is-5'>Select an option from the dropdown</h5>
       <p className='subtitle is-5'>
-        Option: <strong>{JSON.stringify(props.selectedDropdownOption)}</strong>
+        Option: <strong>{JSON.stringify(selectedDropdownOption)}</strong>
       </p>
       <div className='field'>
         <Dropdown
           options={DROPDOWN_TEST_DATA}
           labelKey='label'
-          selectedOptionLabel={props.selectedDropdownOption.label}
+          selectedOptionLabel={selectedDropdownOption.label}
           wrapperClass='normal-width'
-          dispatchHandler={props.selectOption}
+          dispatchHandler={selectOption}
         />
       </div>
     </div>
@@ -56,10 +62,10 @@ const Form: React.FC<FormProps> = props => {
       <h3 className='title is-4'>Checkbox</h3>
       <h5 className='subtitle is-5'>Toggle the checkbox</h5>
       <p className='subtitle is-5'>
-        Checked: <strong>{props.checkboxValue.toString()}</strong>
+        Checked: <strong>{checkboxValue.toString()}</strong>
       </p>
       <div className='field'>
-        <Checkbox onCheck={props.doCheck} checked={props.checkboxValue} />
+        <Checkbox onCheck={doCheck} checked={checkboxValue} />
       </div>
     </div>
   );
